refactor: move Config from agent to config package
